Add fighting system and health

+++ Assets/Source/Scripts/Interface/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Interface
 {
     public interface IDamageable
     {
-        void AcceptDamage(int damage);
+        void AcceptDamage(float damage);
     }
 }
